move routing pre-hook to router.js
