+++ src/main.js
@@ -19,23 +19,6 @@ Vue.material.registerTheme({
   },
 });
 
-// pre-route auth checking
-router.beforeEach((to, from, next) => {
-  const isAuthenticated = store.getters['authentication/isAuthenticated'];
-
-  // check if any of the matched routes require authentication
-  if (to.matched.some(record => record.meta.requiresAuthentication)) {
-    if (isAuthenticated) return next();
-
-    // otherwise, redirect to login, preserving the requested route
-    const { name, fullPath } = to;
-    const query = (name !== undefined) ? { redirect: name } : { prev: fullPath };
-    return next({ name: 'login', query });
-  }
-
-  return next();
-});
-
 // initialize the application store
 store.dispatch('authentication/initialize')
 .then(() => {
